Added authentication requriements to create, edit, and delete a course and for students to add and remove courses from their schedule.

+++ app.js
@@ -14,7 +14,7 @@ app.use(express.json())
 const router = express.Router()
 const secret = "supersecret"
 
-//creating a new Student Account
+//creating a new User Account
 router.post("/user", async(req,res) => {
   if(!req.body.username || !req.body.password || !req.body.role ){
     res.status(400).json({error: "Mssing username, password or role"})
@@ -44,6 +44,7 @@ router.post("/user", async(req,res) => {
   }
 })
 
+//Route to authenticate a user when signing in
 router.post("/auth", async(req,res) => {
   if(!req.body.username || !req.body.password) {
     res.status(400).json({error: "Mssing username or password"})
@@ -107,6 +108,110 @@ router.get("/status", async(req,res) => {
 })
 
 
+//allow only students to add courses to schedule
+router.post('/add-course/:id', async(req,res) => {
+  const token = req.headers["x-auth"];
+  const courseId = req.params.id;
+
+  if (!token) {
+    return res.status(401).json({error: "Missing Token"})
+  }
+
+  try {
+    const decoded = jwt.decode(token, secret);
+    const user = await User.findOne({ username: decoded.username});
+
+    if (!user) {
+      return res.status(404).json({error: "User not found"})
+    } else if (user.role != "student") {
+      return res.status(403).json({error: "Only students can add courses to schedule"})
+    }
+
+    //check if course exists
+    const course = await Course.findById(courseId)
+    if (!course) {
+      return res.status(404).json({error: "Course not found"})
+    }
+
+    //check if already enrolled
+    if (user.enrolledCourses.includes(courseId)) {
+      return res.status(400).json({error: "Already enrolled in this course"})
+    }
+
+    // Add course to enrolled list
+    user.enrolledCourses.push(courseId);
+    await user.save();
+    res.json(user.enrolledCourses);
+  } catch (err) {
+    res.status(401).json({ error: "Invalid token", detail: err.message });
+  }
+})
+
+//allow students to remove a course from their schedule
+router.post('/delete-course/:id', async(req,res) => {
+  const token = req.headers["x-auth"];
+  const courseId = req.params.id;
+
+  if (!token) {
+    return res.status(401).json({error: "Missing Token"})
+  }
+
+  try {
+    const decoded = jwt.decode(token, secret);
+    const user = await User.findOne({ username: decoded.username});
+
+    //check if user exists
+    if (!user) {
+      return res.status(404).json({error: "User not found"})
+      //check if user is a student, if not return 403 forbidden
+    } else if (user.role != "student") {
+      return res.status(403).json({error: "Only students have a course schedule"})
+    }
+
+    
+    // Update students enrolled courses
+    const updatedUser = await User.findOneAndUpdate(
+      { username: decoded.username },
+      { $pull: { enrolledCourses: courseId } },
+      //get the newly updated docuemnt
+      { new: true }
+    ).populate('enrolledCourses');
+
+    res.json(updatedUser.enrolledCourses);
+  } catch (err) {
+    res.status(401).json({ error: "Invalid token", detail: err.message });
+  }
+})
+
+// display all courses for a student, MUST BE SIGNED IN
+router.get('/schedule', async(req,res) => {
+  const token = req.headers["x-auth"];
+
+  if (!token) {
+    return res.status(401).json({error: "Missing Token"})
+  }
+
+  try {
+    const decoded = jwt.decode(token, secret);
+    //locate the user in the database and populate the enrolledCourses property with the full course object
+    const user = await User.findOne({ username: decoded.username}).populate("enrolledCourses");
+
+    //check if user exists
+    if (!user) {
+      return res.status(404).json({error: "User not found"})
+      //check if user is a teacher
+    } else if (user.role != "student") {
+      return res.status(403).json({error: "Only students have a course schedule"})
+    }
+    
+    //send frontend json data of all enrolled courses and their properties
+    res.json(user.enrolledCourses);
+  } catch (err) {
+    res.status(401).json({ error: "Invalid token", detail: err.message });
+  }
+})
+
+
 // Get list of all courses in the database... GET Request to https://group-5-final-project-backend.onrender.com/api/courses/
 router.get("/courses", async(req,res) => {
   try{
@@ -130,7 +235,26 @@ router.get("/courses/:id", async(req,res) => {
 
 //Add a course to the database... POST Request to https://group-5-final-project-backend.onrender.com/api/courses
 router.post("/courses", async(req,res) => {
+  const token = req.headers["x-auth"];
+
+  if (!token) {
+    return res.status(401).json({error: "Missing Token"})
+  }
+
   try {
+    const decoded = jwt.decode(token, secret);
+    //locate the user in the database and populate the enrolledCourses property with the full course object
+    const user = await User.findOne({ username: decoded.username});
+    
+    //check if user exists
+    if (!user) {
+      return res.status(404).json({error: "User not found"})
+      //check if user is teacher, if not then return 403 forbidden
+    } else if (user.role != "teacher") {
+      return res.status(403).json({error: "Only teachers can create a course"})
+    }
+
+
     const course = await new Course(req.body)
     await course.save()
     res.status(201).json(course)
@@ -141,7 +265,27 @@ router.post("/courses", async(req,res) => {
 
 //update a course in the database.. PUT Request to https://group-5-final-project-backend.onrender.com/api/courses/:id
 router.put("/courses/:id", async(req,res) => {
-  try{
+  const token = req.headers["x-auth"];
+
+  if (!token) {
+    return res.status(401).json({error: "Missing Token"})
+  }
+
+  
+
+  try {
+    const decoded = jwt.decode(token, secret);
+    //locate the user in the database and populate the enrolledCourses property with the full course object
+    const user = await User.findOne({ username: decoded.username});
+    
+    //check if user exists
+    if (!user) {
+      return res.status(404).json({error: "User not found"})
+      //check if user is teacher, if not then return 403 forbidden
+    } else if (user.role != "teacher") {
+      return res.status(403).json({error: "Only teachers can create a course"})
+    }
+
     const course = req.body
     await Course.updateOne({_id: req.params.id}, course)
     res.sendStatus(204)
@@ -152,7 +296,25 @@ router.put("/courses/:id", async(req,res) => {
 
 //delete a course in the database... DELETE Request to https://group-5-final-project-backend.onrender.com/api/courses/:id
 router.delete("/courses/:id", async(req,res) => {
-  try{
+  const token = req.headers["x-auth"];
+
+  if (!token) {
+    return res.status(401).json({error: "Missing Token"})
+  }
+
+  try {
+    const decoded = jwt.decode(token, secret);
+    //locate the user in the database and populate the enrolledCourses property with the full course object
+    const user = await User.findOne({ username: decoded.username});
+    
+    //check if user exists
+    if (!user) {
+      return res.status(404).json({error: "User not found"})
+      //check if user is teacher, if not then return 403 forbidden
+    } else if (user.role != "teacher") {
+      return res.status(403).json({error: "Only teachers can create a course"})
+    }
+
     await Course.deleteOne({_id: req.params.id})
     res.sendStatus(204)
   }catch (err) {
